量化/requirements.txt
ccxt>=4.2.15
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0